omnisearch-backend/requirements.txt
fastapi
uvicorn
python-dotenv
groq
pinecone-client
cohere-ai
google-generativeai
pydantic
python-multipart